myAn/requirements.txt
Django==1.11
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.2.1
numpy==1.18.1
openpyxl==2.6.4
pandas==0.25.3
pycountry==19.8.18
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
xlrd==1.2.0
